src/scraper.py
import requests
from bs4 import BeautifulSoup
from newspaper import Article
from typing import List
import logging
from datetime import datetime
from .models import NewsArticle
from config import USER_AGENT

class NewsScraper:
    def __init__(self):
        self.headers = {'User-Agent': USER_AGENT}
        self.logger = logging.getLogger(__name__)

    def get_news_links(self, url: str) -> List[str]:
        """특정 사이트에서 뉴스 링크들을 수집합니다."""
        try:
            response = requests.get(url, headers=self.headers)
            soup = BeautifulSoup(response.text, 'html.parser')
            # 실제 구현시 사이트별 링크 추출 로직 구현 필요
            links = []  # 뉴스 링크 추출 로직
            return links
        except Exception as e:
            self.logger.error(f"링크 수집 중 오류 발생: {str(e)}")
            return []

    def parse_article(self, url: str) -> NewsArticle:
        """개별 뉴스 기사를 파싱합니다."""
        try:
            article = Article(url)
            article.download()
            article.parse()
            
            # 기사 정보 추출
            title = article.title
            content = article.text
            published_date = article.publish_date or datetime.now()
            
            # 추가 정보 추출 로직 구현 필요
            company_name = self._extract_company_name(content)
            stock_code = self._get_stock_code(company_name)
            publisher = self._extract_publisher(article)
            daily_occurrence = self._count_daily_occurrences(title, published_date)
            
            return NewsArticle(
                title=title,
                company_name=company_name,
                stock_code=stock_code,
                content=content,
                price_impact=0.0,  # 분석기에서 설정될 값
                published_date=published_date,
                publisher=publisher,
                daily_occurrence=daily_occurrence,
                url=url
            )
        except Exception as e:
            self.logger.error(f"기사 파싱 중 오류 발생: {str(e)}")
            raise

    def _extract_company_name(self, content: str) -> str:
        """기사 내용에서 회사명을 추출합니다."""
        # 구현 필요
        return ""

    def _get_stock_code(self, company_name: str) -> str:
        """회사명으로 주식 코드를 조회합니다."""
        # 구현 필요
        return ""

    def _extract_publisher(self, article: Article) -> str:
        """기사의 언론사를 추출합니다."""
        # 구현 필요
        return ""

    def _count_daily_occurrences(self, title: str, date: datetime) -> int:
        """같은 날짜에 유사한 기사가 몇 번 등장했는지 계산합니다."""
        # 구현 필요
        return 1 