requirements.txt
beautifulsoup4==4.12.3
requests==2.31.0
pandas==2.2.1
pymongo==4.6.1
python-dotenv==1.0.1
newspaper3k==0.2.8 